Make "comms" module
