+++ source/kernel/src/comms/mod.rs
@@ -1,0 +1,2 @@
+pub mod bbq;
+pub mod kchannel;
