+++ source/kernel/src/drivers/mod.rs
@@ -1,0 +1,1 @@
+pub mod serial_mux;
