+++ source/kernel/src/drivers/serial_mux.rs
@@ -1,0 +1,5 @@
+use crate::comms::bbq::BBQBidiHandle;
+use spitebuf::MpMcQueue;
+pub struct SerialMux {
+    serial_port: BBQBidiHandle,
+}
