+++ source/kernel/src/lib.rs
@@ -1,7 +1,8 @@
 #![no_std]
 #![allow(clippy::missing_safety_doc)]
 
-pub mod bbq;
+pub mod comms;
+pub mod drivers;
 pub(crate) mod fmt;
 
 use abi::{
@@ -11,6 +12,7 @@ use abi::{
     },
     syscall::{DriverKind, KernelResponse, UserRequest},
 };
+use comms::kchannel::KChannel;
 use core::{
     cell::UnsafeCell,
     ops::{Deref, DerefMut},
@@ -23,8 +25,8 @@ use maitake::{
     task::Storage,
 };
 use mnemos_alloc::{
-    containers::{HeapArc, HeapArray, HeapBox, HeapFixedVec},
-    heap::{AHeap, HeapGuard},
+    containers::{HeapBox, HeapFixedVec},
+    heap::AHeap,
 };
 use tracing::info;
 
@@ -300,53 +302,5 @@ impl<F: Future + 'static> Storage<&'static StaticScheduler, F> for HBStorage {
 
     fn from_raw(ptr: NonNull<MaitakeTask<&'static StaticScheduler, F, Self>>) -> HeapBox<Task<F>> {
         unsafe { HeapBox::from_leaked(ptr.cast::<Task<F>>()) }
-    }
-}
-
-use spitebuf::MpMcQueue;
-
-mod sealed {
-    use super::*;
-
-    pub struct SpiteData<T> {
-        pub(crate) data: HeapArray<UnsafeCell<spitebuf::Cell<T>>>,
-    }
-
-    unsafe impl<T: Sized> spitebuf::Storage<T> for SpiteData<T> {
-        fn buf(&self) -> (*const UnsafeCell<spitebuf::Cell<T>>, usize) {
-            let ptr = self.data.as_ptr();
-            let len = self.data.len();
-            (ptr, len)
-        }
-    }
-}
-
-pub struct KChannel<T> {
-    q: HeapArc<MpMcQueue<T, sealed::SpiteData<T>>>,
-}
-
-impl<T> Clone for KChannel<T> {
-    fn clone(&self) -> Self {
-        Self { q: self.q.clone() }
-    }
-}
-
-impl<T> Deref for KChannel<T> {
-    type Target = MpMcQueue<T, sealed::SpiteData<T>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.q
-    }
-}
-
-impl<T> KChannel<T> {
-    pub fn new(guard: &mut HeapGuard, count: usize) -> Self {
-        let func = || UnsafeCell::new(spitebuf::single_cell::<T>());
-
-        let ba = guard.alloc_box_array_with(func, count).unwrap();
-        let q = MpMcQueue::new(sealed::SpiteData { data: ba });
-        Self {
-            q: guard.alloc_arc(q).map_err(drop).unwrap(),
-        }
     }
 }
